// src/generated/models/Cfm_insurancecompaniesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cfm_insurancecompaniescfm_companytype = {
  766340000: 'Africa',
  766340001: 'Cash',
  766340002: 'Government ',
  766340003: 'Corporate ',
  766340004: 'Insurance ',
  766340005: 'Banks',
  766340006: 'Others',
  766340007: 'Petroleum'
} as const;
export type Cfm_insurancecompaniescfm_companytype = keyof typeof Cfm_insurancecompaniescfm_companytype;
export const Cfm_insurancecompaniescfm_customerclass = {
  766340000: 'A+',
  766340001: 'A',
  766340002: 'B',
  766340003: 'B+',
  766340004: 'C+',
  766340005: 'C'
} as const;
export type Cfm_insurancecompaniescfm_customerclass = keyof typeof Cfm_insurancecompaniescfm_customerclass;
export const Cfm_insurancecompaniesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cfm_insurancecompaniesstatecode = keyof typeof Cfm_insurancecompaniesstatecode;
export const Cfm_insurancecompaniesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cfm_insurancecompaniesstatuscode = keyof typeof Cfm_insurancecompaniesstatuscode;

export interface Cfm_insurancecompaniesBase {
  "cfm_BU@odata.bind": string;
  cfm_category?: string;
  cfm_code: string;
  cfm_companyname?: string;
  cfm_companytype?: Cfm_insurancecompaniescfm_companytype;
  cfm_customerclass?: Cfm_insurancecompaniescfm_customerclass;
  cfm_insurancecompanyid: string;
  cfm_name: string;
  "cfm_PaymentTerm@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cfm_insurancecompaniesstatecode;
  statuscode?: Cfm_insurancecompaniesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cfm_insurancecompanies extends Cfm_insurancecompaniesBase {
  cfm_buname?: string;
  cfm_butext?: string;
  cfm_companytypename?: string;
  cfm_customerclassname?: string;
  cfm_paymenttermname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cfm_bu?: object;
  _cfm_bu_value?: string;
  cfm_paymentterm?: object;
  _cfm_paymentterm_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}